Move configs to separate file.
